Kernel: Xiaomi kernel changes for Redmi 4/4X
The Patch based on QualComm release TAG:LA.UM.5.6.r1-04600-89xx.0-1
The kernel config file used is S88536AA2-perf_defconfig

+++ kernel/sched/sched_avg.c
@@ -107,7 +107,7 @@ EXPORT_SYMBOL(sched_get_nr_running_avg);
  */
 void sched_update_nr_prod(int cpu, long delta, bool inc)
 {
-	int diff;
+	s64 diff;
 	s64 curr_time;
 	unsigned long flags, nr_running;
